--- .changeset/plop-create-side-bar.md
@@ -1,5 +1,0 @@
----
-'@ag.ds-next/react': minor
----
-
-Created new component `SideBar`
