--- packages/react/side-bar/package.json
@@ -1,3 +1,0 @@
-{
-	"main": "dist/ag.ds-next-react-side-bar.cjs.js"
-}
