+++ packages/react/src/_collapsing-side-bar/CollapsingSideBar.test.tsx
@@ -1,21 +1,21 @@
 import '@testing-library/jest-dom';
 import 'html-validate/jest';
 import { cleanup, render } from '../../../../test-utils';
-import { SideBar } from './SideBar';
-import type { SideBarProps } from './SideBar';
+import { CollapsingSideBar } from './CollapsingSideBar';
+import type { SideBarProps } from './CollapsingSideBar';
 
 afterEach(cleanup);
 
 const defaultSideBarProps: SideBarProps = {
-	children: 'SideBar Content',
+	children: 'CollapsingSideBar Content',
 	collapseButtonLabel: 'Collapse button label',
 };
 
 function renderSideBar(props: SideBarProps) {
-	return render(<SideBar {...props} />);
+	return render(<CollapsingSideBar {...props} />);
 }
 
-describe('SideBar', () => {
+describe('CollapsingSideBar', () => {
 	it('renders correctly', () => {
 		const { container } = renderSideBar(defaultSideBarProps);
 		expect(container).toMatchSnapshot();
