--- packages/react/src/side-bar/docs/code.mdx
@@ -1,5 +1,0 @@
-## Props
-
-<ComponentPropsTable name="SideBar" />
-
-<ComponentPropsTable name="SideBarTitle" />
