--- packages/react/src/side-bar/docs/overview.mdx
@@ -1,51 +1,0 @@
----
-title: Side bar
-description: A layout component which frames secondary navigation content.
-group: Layout
-storybookPath: /story/Layout-SideBar--basic
-relatedComponents: [side-nav, progress-indicator]
----
-
-Side bar is a layout component which frames secondary navigation content. It's collapses on mobile to allow the user to focus on the primary content.
-
-Side bar is used internally by the [SideNav](/components/side-nav) and [ProgressIndicator](/components/progress-indicator) components. You should consider using those components before using Side bar directly.
-
-```jsx live
-<SideBar collapseButtonLabel="Collapse button label">
-	<SideBarTitle title="Side bar title" />
-	<Box paddingY={1}>Side bar Content area</Box>
-</SideBar>
-```
-
-<DoHeading />
-
-- Use Side bar to contain data filters.
-- Use the SideBarTitle component to add a title above the content area.
-
-<DontHeading />
-
-- Compose custom navigation components without considering if [SideNav](/components/side-nav) or [ProgressIndicator](/components/progress-indicator) is right for your usecase.
-
-## Collapse button label
-
-For narrower screens, the SideBar collapses into an accordion. The collapse button label is used to describe the content of the SideBar to screen readers.
-
-Navigation components use the label 'In this section'. Consider what is the most appropriate label for your usecase.
-
-```jsx live
-<SideBar collapseButtonLabel="1 of 5 steps completed">
-	<SideBarTitle title="Side bar title" />
-	<Box paddingY={1}>Side bar Content area</Box>
-</SideBar>
-```
-
-## Side bar title
-
-The SideBarTitle component is used to add a title above the content area.
-
-```jsx live
-<SideBar collapseButtonLabel="1 of 5 steps completed">
-	<SideBarTitle title="Progress" subtitle="Step 1 of 5" />
-	<Box paddingY={1}>Side bar Content area</Box>
-</SideBar>
-```
