--- packages/react/src/side-bar/index.ts
@@ -1,3 +1,0 @@
-export * from './SideBar';
-export * from './SideBarTitle';
-export * from './utils';
